# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Document Recognition API",
        "description": "API для распознавания документов через внешний сервис bescan.ru",
        "contact": {
            "name": "API Support",
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8080/api/v1",
            "description": "Local Development Server"
        }
    ],
    "paths": {
        "/analytics": {
            "get": {
                "tags": [
                    "Analytics"
                ],
                "summary": "Получить аналитику распознавания документов",
                "description": "Возвращает статистику по распознаванию документов за указанный период",
                "operationId": "getAnalytics",
                "parameters": [
                    {
                        "name": "start_date",
                        "in": "query",
                        "description": "Начальная дата периода (формат: Y-m-d)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2024-01-01"
                        }
                    },
                    {
                        "name": "end_date",
                        "in": "query",
                        "description": "Конечная дата периода (формат: Y-m-d)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2024-12-31"
                        }
                    },
                    {
                        "name": "document_type",
                        "in": "query",
                        "description": "Фильтр по типу документа",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "PASSPORT",
                                "PASSPORT_REG",
                                "DLIC",
                                "SNILS",
                                "STS"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Аналитика получена успешно",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "total_requests": {
                                                    "description": "Общее количество запросов",
                                                    "type": "integer",
                                                    "example": 1000
                                                },
                                                "successful_requests": {
                                                    "description": "Успешных запросов",
                                                    "type": "integer",
                                                    "example": 950
                                                },
                                                "failed_requests": {
                                                    "description": "Неудачных запросов",
                                                    "type": "integer",
                                                    "example": 50
                                                },
                                                "success_rate": {
                                                    "description": "Процент успешных запросов",
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 95
                                                },
                                                "average_processing_time": {
                                                    "description": "Среднее время обработки в секундах",
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 2.5
                                                },
                                                "requests_by_type": {
                                                    "description": "Запросы по типам документов",
                                                    "type": "object",
                                                    "example": {
                                                        "PASSPORT": 500,
                                                        "DLIC": 300
                                                    }
                                                },
                                                "requests_by_status": {
                                                    "description": "Запросы по статусам",
                                                    "type": "object",
                                                    "example": {
                                                        "completed": 950,
                                                        "failed": 50
                                                    }
                                                },
                                                "daily_stats": {
                                                    "description": "Ежедневная статистика",
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Доступ запрещен",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Access denied."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
...
}